services/auth/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.9.0
